refactor(cluster)!: Route K-means natively
Keep arithmetic in the concrete f32 image precision while normalized initialization weights and shifted means prevent finite-intermediate overflow.

BREAKING CHANGE: KMeansSegmentation construction, configuration, and application are fallible; configuration fields are private and CLI K-means requires native image formats.

+++ crates/ritk-cli/src/commands/segment/args.rs
@@ -103,7 +103,7 @@ pub struct SegmentArgs {
 
     /// Centroid-displacement convergence tolerance for `kmeans`. Default: 1e-6.
     #[arg(long, value_name = "FLOAT")]
-    pub kmeans_tolerance: Option<f64>,
+    pub kmeans_tolerance: Option<f32>,
 
     /// Deterministic seed for k-means++ initialization. Default: 42.
     #[arg(long, value_name = "INT")]
